Validation and clear button style improvements
